Add config parsing and authorization headers
This commit adds a simple config parser. The idea is to hold secrets
and other information that might not fit as a flag in a config.ini
file. We will use this to hold a bearer token which will be used for
authorization. This token is constant and the authorization logic
blocks all endpoints. This is good enough for now, even though it
makes the dashboard practically useless.